[CPU] Renamed common->plugin
