+++ src/plugins/intel_cpu/src/emitters/snippets/x64/cpu_generator.cpp
@@ -14,10 +14,10 @@
 #include "emitters/snippets/x64/jit_snippets_emitters.hpp"
 #include "emitters/snippets/x64/jit_fill_emitter.hpp"
 #include "emitters/snippets/x64/jit_horizon_emitter.hpp"
-#include "emitters/common/x64/jit_eltwise_emitters.hpp"
-#include "emitters/common/x64/jit_dnnl_emitters.hpp"
-#include "emitters/common/x64/jit_dnnl_ext_emitters.hpp"
-#include "emitters/common/x64/jit_conversion_emitters.hpp"
+#include "emitters/plugin/x64/jit_eltwise_emitters.hpp"
+#include "emitters/plugin/x64/jit_dnnl_emitters.hpp"
+#include "emitters/plugin/x64/jit_dnnl_ext_emitters.hpp"
+#include "emitters/plugin/x64/jit_conversion_emitters.hpp"
 
 #include "transformations/snippets/x64/op/load_convert.hpp"
 #include "transformations/snippets/x64/op/store_convert.hpp"
