+++ src/plugins/intel_cpu/src/emitters/snippets/x64/jit_kernel_emitter.hpp
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "emitters/common/x64/jit_emitter.hpp"
+#include "emitters/plugin/x64/jit_emitter.hpp"
 
 #include "jit_container_emitter.hpp"
