+++ src/plugins/intel_cpu/src/nodes/eltwise.cpp
@@ -11,9 +11,9 @@
 #include "cpu/x64/injectors/jit_uni_quantization_injector.hpp"
 #include "cpu_types.h"
 #include "dnnl_extension_utils.h"
-#include "emitters/common/x64/jit_bf16_emitters.hpp"
-#include "emitters/common/x64/jit_dnnl_emitters.hpp"
-#include "emitters/common/x64/jit_eltwise_emitters.hpp"
+#include "emitters/plugin/x64/jit_bf16_emitters.hpp"
+#include "emitters/plugin/x64/jit_dnnl_emitters.hpp"
+#include "emitters/plugin/x64/jit_eltwise_emitters.hpp"
 #include "fake_quantize.h"
 #include "input.h"
 #include "memory_desc/dnnl_blocked_memory_desc.h"
