+++ src/plugins/intel_cpu/src/nodes/kernels/x64/jit_kernel.hpp
@@ -4,7 +4,7 @@
 
 #pragma once
 #include "cpu/x64/jit_generator.hpp"
-#include "emitters/common/x64/jit_load_store_emitters.hpp"
+#include "emitters/plugin/x64/jit_load_store_emitters.hpp"
 #include <common/nstl.hpp>
 #include <type_traits>
 #include <functional>
