+++ src/plugins/intel_cpu/src/nodes/mha.cpp
@@ -7,8 +7,8 @@
 #include "common/cpu_memcpy.h"
 #include "cpu/x64/jit_generator.hpp"
 #include "dnnl_extension_utils.h"
-#include "emitters/common/x64/jit_dnnl_emitters.hpp"
-#include "emitters/common/x64/jit_load_store_emitters.hpp"
+#include "emitters/plugin/x64/jit_dnnl_emitters.hpp"
+#include "emitters/plugin/x64/jit_load_store_emitters.hpp"
 #include "openvino/core/parallel.hpp"
 #include "openvino/opsets/opset1.hpp"
 #include "transformations/cpu_opset/x64/op/mha.hpp"
