+++ src/plugins/intel_cpu/src/nodes/region_yolo.cpp
@@ -13,7 +13,7 @@
 #include "common/cpu_convert.h"
 #include "cpu/x64/jit_generator.hpp"
 #include "cpu/x64/injectors/jit_uni_eltwise_injector.hpp"
-#include "emitters/common/x64/jit_bf16_emitters.hpp"
+#include "emitters/plugin/x64/jit_bf16_emitters.hpp"
 #include "utils/bfloat16.hpp"
 
 using namespace dnnl::impl::cpu;
